requirements.txt
Flask>=2.3
psycopg2-binary>=2.9